project 13 - add a cabin duplication button

+++ react in study Jonas Schmedtmann/17-the-wild-oasis/starter/src/services/apiCabins.js
@@ -53,6 +53,8 @@ export async function createEditCabin(newCabin, idCabin) {
     throw new Error("Cabin could not be insert");
   }
 
+  if (hasImage) return data;
+
   const { error: storageError } = await supabase.storage.from("cabin-images").upload(imageName, newCabin.image);
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
